pyproject.toml
[project]
name = "clawdface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "livekit-agents[trugen]>=1.4",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
    "websockets>=10.0,<16.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
